Face Similarity/requirements.txt
face-recognition==1.3.0
Flask==2.2.2
numpy==1.19.5
Pillow==9.2.0
requests==2.28.1